fixed warnings

+++ dchu.cpp
@@ -606,7 +606,10 @@ void Enemy::initEnemy(World* w, float mag = 48.0f)
         DWARN("Error, World Not Found\n");
     }
     auto cells = w->cells;
-    v2f w_max = {cells.size(), cells[0].size()};
+    v2f w_max = {
+        static_cast<float>cells.size(),
+        static_cast<float>cells[0].size()
+    };
     w_max[0] *= m_mag;
     w_max[1] *= m_mag;
     auto [health, collide, sprite, transform, physics, navigate] =
